Remove bias from the die throw; change randomness range
The old code was generating random 32-bit numbers, or anything in the range of:

    0 <= n <= 4294967295

Unfortunately, 2^32 is not divisible by 6, which is how many faces are on a
die. As such, 1-4 will be tossed on a die more frequently than 5 & 6.

To remove this bias, a range needs to be created that is a multiple of 6. For
simplicity, rather than using a 32-bit random number, this code changes it to a
more manageable random 8-bit number, then adjusts the range to:

    0 <= n <= 251

This has a total of 252 elements, which is 42*6. As such, each number will be
equally likely to appear, removing the prior bias.