Add subresource integrity.
SRI is a W3C recommendation to ensure that clients requesting subresources are
getting the data correctly. The specification can be read at
https://www.w3.org/TR/SRI/. Browser support is limited, with about 70% support.
However, browsers that do not support SRI should silently ingore it.

This commit adds SRI to the projects hosted CSS and JS using SHA-256. This does
remove the ability to run the HTML "offline". Instead, the code will be needed
to be served out of a web server, even if the webserver is 127.0.0.1.